recruit_final_v2 7 copy/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
google-generativeai>=0.5.0
imap-tools>=1.6.0
python-multipart>=0.0.9
pypdf>=4.0.0
mammoth>=1.6.0
python-dotenv>=1.0.0
pinecone>=3.0.0
numpy>=1.26.0
jinja2>=3.1.3
websockets>=12.0
aiofiles>=23.2.1
google-genai>=0.8.0
httpx>=0.27.0
